feat: extract hooks to make route for tasks #5305
